chore: restructured api folder
