+++ src/api/settings/Profile.ts
@@ -1,8 +1,34 @@
 import { useUser } from '@/context/AppContext';
-import { EditProfile, MutationErrorPayload } from '@/types/types';
-import { useState } from 'react';
+import { MutationErrorPayload } from '@/types/types';
 import { toast } from 'sonner';
-import { useGetProfile } from './getProfile';
+import { EditProfile } from '@/types/types';
+import { useState } from 'react';
+
+export const useGetProfile = () => {
+  const { setUser } = useUser();
+  const getProfile = async (token: string) => {
+    try {
+      const data = await fetch(
+        `${import.meta.env.VITE_API_BASE_URL}admin/settings/profile`,
+        {
+          method: 'GET',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            Accept: 'application/json',
+          },
+        },
+      );
+      const res = await data.json();
+      if (res?.error) {
+        toast.error(res?.message);
+      } else setUser(res?.admin_data);
+    } catch (err) {
+      toast.error((err as MutationErrorPayload)?.data?.message);
+    }
+  };
+
+  return { getProfile };
+};
 
 export const useEditProfile = () => {
   const { user, token } = useUser();
